refactor: move JSON functionality to `common` module
* `getStringOrNull`

Removes a cyclic dependency: libAnki <-> AnkiDroid

Issue 18015
