+++ AnkiDroid/src/main/java/com/ichi2/libanki/Field.kt
@@ -17,7 +17,7 @@
 package com.ichi2.libanki
 
 import com.ichi2.anki.common.json.JSONObjectHolder
-import com.ichi2.anki.utils.ext.getStringOrNull
+import com.ichi2.anki.common.utils.ext.getStringOrNull
 import net.ankiweb.rsdroid.RustCleanup
 import org.json.JSONObject
